pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pefile>=2024.8.26",
    "psycopg2-binary>=2.9.10",
    "python-magic>=0.4.27",
    "twilio>=9.4.6",
    "werkzeug>=3.1.3",
    "yara-python>=4.5.1",
    "flask-wtf>=1.2.2",
    "oauthlib>=3.2.2",
    "openai>=1.64.0",
    "anthropic>=0.46.0",
    "slack-sdk>=3.34.0",
    "sendgrid>=6.11.0",
    "trafilatura>=2.0.0",
]
